+++ truffle/contracts/CryptographyWrapper.sol
@@ -4,7 +4,7 @@ import "./Cryptography.sol";
 
 contract CryptographyWrapper {
 
-    function alphaData(address from, address to, uint256 value, uint t0, uint t1) public pure returns (bytes32) {
+    function alphaData(address from, address to, uint256 value, uint256 t0, uint256 t1) public pure returns (bytes32) {
         return Cryptography.alphaData(from, to, value, t0, t1);
     }
